Minor edits to privacy policy
This PR provides minor edits to spelling and verbiage for consistency, and hyperlinks where missing.

Question about this paragraph:
**What information do we collect?**

Pre- and post-workshop surveys represent a formal part of the workshop training process and 
are vital to ensure that our lessons, organisation, Code of Conduct, and workshops adhere 
to our vision, mission, and standards. The information collected within these surveys is 
anonymised and used in reporting and impact studies.

Should it be vision, mission, and values (instead of standards)?