Update etherpads.md
Merged duplicate segments talking about  "list of our most commonly used pads is found at ..." from line 5 and 15.

Added link to the CodiMD at the bottom as a suggested alternative. If anyone reached to the bottom of the page, that might mean that the visitor would like to explore more.