Update codimd.md
Corrected link to the CodiMD.

Added link to the Etherpad at the bottom as a suggested alternative. If anyone reached to the bottom of the page, that might mean that the visitor would like to explore more. Not all visitors would be reading the webpage sequentially starting from Etherpad page and may have reached the CodiMD page directly.